+++ Base.lproj/CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Build 3
 * Added EULA popup to the first time the user logins
 * Changed report behavior for channels, videos, and comments
+### Build 4
+* Added EULA to additional settings
 ## v0.7 (Beta Release)
 * This is the beta release of BlurrMC. The beta release includes all the software neccessary for beta testers to start testing all the functionality of the app. Any future revisions up to v1.0 will be considered a beta.
 ## v0.5 (ALPHA)